moved the models into their own files
