+++ back_end/models/__init__.py
@@ -1,0 +1,25 @@
+from .user import (
+    UserBase,
+    UserResponse,
+    UserCreate,
+    UserRegister,
+    UserUpdate,
+    UserUpdateMe,
+    UpdatePassword,
+    User,
+    UserPublic,
+    UsersPublic,
+)
+
+__all__ = [
+    "UserBase",
+    "UserResponse",
+    "UserCreate",
+    "UserRegister",
+    "UserUpdate",
+    "UserUpdateMe",
+    "UpdatePassword",
+    "User",
+    "UserPublic",
+    "UsersPublic",
+]
